Use -Werror=format-security for BioC >= 3.21 + R >= 4.5
The R-fix-flags.sh script now adds the -Werror=format-security flag (in
addition to -Wall) to the *FLAGS compiler variables. It also asks the
user to confirm that the script is being used for BioC >= 3.21 + R >= 4.5
before proceeding.

For BioC < 3.21 + R < 4.5, R-fix-flags-old.sh should be used instead of
R-fix-flags.sh. The R-fix-flags-old.sh script implements the old behavior
i.e. it only adds the -Wall flag to the *FLAGS compiler variables.